backend/requirements.production.txt
# Production requirements for VittCott Trading Platform

# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# WebSocket support
websockets==12.0

# Data processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
yfinance==0.2.18

# Database and caching
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
PyJWT==2.8.0
bcrypt==4.0.1
python-multipart==0.0.6

# HTTP client for external APIs
aiohttp==3.9.1
httpx==0.25.2

# System monitoring
psutil==5.9.6

# Text processing
textblob==0.17.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring and observability
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Rate limiting
slowapi==0.1.9

# CORS support
fastapi-cors==0.0.6

# Background tasks
celery==5.3.4

# Logging
structlog==23.2.0

# Testing (optional for production)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools (optional for production)
black==23.11.0
flake8==6.1.0
mypy==1.7.1
